refactor: 클래스 리패키징
- Event 관련 엔티티 패키지 이동
- Repository 패키지 이동
