+++ src/main/java/com/softeer/podo/admin/exception/EventNotFoundException.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.admin.model.exception;
+package com.softeer.podo.admin.exception;
 
 public class EventNotFoundException extends RuntimeException {
 	private String message;
