+++ src/main/java/com/softeer/podo/admin/model/dto/ArrivalUserDto.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.admin.model.dto.user;
+package com.softeer.podo.admin.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
@@ -7,10 +7,11 @@ import lombok.Data;
 
 @Data
 @Builder
-public class LotsUserDto {
+public class ArrivalUserDto {
 	private Long id;
 	private String name;
 	private String phoneNum;
+	private int rank;
 	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
 	private LocalDateTime createdAt;
 	private String reward;
