+++ src/main/java/com/softeer/podo/admin/model/dto/UserListDto.java
@@ -1,0 +1,4 @@
+package com.softeer.podo.admin.model.dto;
+
+public interface UserListDto {
+}
