--- src/main/java/com/softeer/podo/admin/model/dto/user/UserListDto.java
@@ -1,4 +1,0 @@
-package com.softeer.podo.admin.model.dto.user;
-
-public interface UserListDto {
-}
