+++ src/main/java/com/softeer/podo/admin/service/AdminService.java
@@ -6,20 +6,20 @@ import com.softeer.podo.admin.model.dto.request.EventConfigRequestDto;
 import com.softeer.podo.admin.model.dto.request.EventRewardConfigRequestDto;
 import com.softeer.podo.admin.model.dto.response.EventListResponseDto;
 import com.softeer.podo.admin.model.dto.response.EventRewardConfigResponseDto;
-import com.softeer.podo.admin.model.dto.user.ArrivalUserDto;
-import com.softeer.podo.admin.model.dto.user.ArrivalUserListDto;
-import com.softeer.podo.admin.mapper.EventMapper;
-import com.softeer.podo.admin.mapper.UserMapper;
-import com.softeer.podo.admin.model.dto.user.LotsUserListDto;
-import com.softeer.podo.admin.model.entity.ArrivalUser;
-import com.softeer.podo.admin.model.entity.Event;
-import com.softeer.podo.admin.model.entity.EventReward;
-import com.softeer.podo.admin.model.exception.EventNotFoundException;
-import com.softeer.podo.admin.repository.EventRepository;
-import com.softeer.podo.admin.model.entity.LotsUser;
-import com.softeer.podo.admin.repository.ArrivalUserRepository;
-import com.softeer.podo.admin.repository.EventRewardRepository;
-import com.softeer.podo.admin.repository.LotsUserRepository;
+import com.softeer.podo.admin.model.dto.ArrivalUserDto;
+import com.softeer.podo.admin.model.dto.ArrivalUserListDto;
+import com.softeer.podo.admin.model.mapper.EventMapper;
+import com.softeer.podo.admin.model.mapper.UserMapper;
+import com.softeer.podo.admin.model.dto.LotsUserListDto;
+import com.softeer.podo.event.model.entity.ArrivalUser;
+import com.softeer.podo.event.model.entity.Event;
+import com.softeer.podo.event.model.entity.EventReward;
+import com.softeer.podo.admin.exception.EventNotFoundException;
+import com.softeer.podo.event.repository.EventRepository;
+import com.softeer.podo.event.model.entity.LotsUser;
+import com.softeer.podo.event.repository.ArrivalUserRepository;
+import com.softeer.podo.event.repository.EventRewardRepository;
+import com.softeer.podo.event.repository.LotsUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
