+++ src/main/java/com/softeer/podo/admin/validation/ValidationGroups.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.admin.model.entity.validation;
+package com.softeer.podo.admin.validation;
 
 
 public class ValidationGroups {
