+++ src/main/java/com/softeer/podo/event/model/entity/EventType.java
@@ -1,4 +1,5 @@
-package com.softeer.podo.admin.model.entity;
+package com.softeer.podo.event.model.entity;
+
 
 import com.softeer.podo.common.entity.DateEntity;
 import jakarta.persistence.*;
@@ -8,19 +9,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "testusers")
+@Table(name = "event_types")
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class TestUser extends DateEntity {
-
+public class EventType extends DateEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name ="testuser_id")
-    private Long id; //고유 pk
-
-    private String name;
-    private String phoneNum;
-    private Role role;
+    @Column(name = "type_id")
+    private Long id;
+    @Column(name = "type")
+    private String type;
 }
