+++ src/main/java/com/softeer/podo/event/model/entity/LotsComment.java
@@ -2,7 +2,6 @@ package com.softeer.podo.event.model.entity;
 
 
 import com.softeer.podo.common.entity.DateEntity;
-import com.softeer.podo.admin.model.entity.LotsUser;
 import jakarta.persistence.*;
 import lombok.*;
