+++ src/main/java/com/softeer/podo/event/model/entity/Role.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.admin.model.entity;
+package com.softeer.podo.event.model.entity;
 
 public enum Role {
     ROLE_USER, ROLE_ADMIN, ROLE_ANONYMOUS
