+++ src/main/java/com/softeer/podo/event/repository/ArrivalUserRepository.java
@@ -1,6 +1,6 @@
-package com.softeer.podo.admin.repository;
+package com.softeer.podo.event.repository;
 
-import com.softeer.podo.admin.model.entity.ArrivalUser;
+import com.softeer.podo.event.model.entity.ArrivalUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArrivalUserRepository extends JpaRepository<ArrivalUser, Long> {
