+++ src/main/java/com/softeer/podo/event/repository/EventRepository.java
@@ -1,6 +1,6 @@
-package com.softeer.podo.admin.repository;
+package com.softeer.podo.event.repository;
 
-import com.softeer.podo.admin.model.entity.Event;
+import com.softeer.podo.event.model.entity.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventRepository extends JpaRepository<Event, Long> {
