+++ src/main/java/com/softeer/podo/event/repository/EventRewardRepository.java
@@ -1,7 +1,7 @@
-package com.softeer.podo.admin.repository;
+package com.softeer.podo.event.repository;
 
-import com.softeer.podo.admin.model.entity.Event;
-import com.softeer.podo.admin.model.entity.EventReward;
+import com.softeer.podo.event.model.entity.Event;
+import com.softeer.podo.event.model.entity.EventReward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
