+++ src/main/java/com/softeer/podo/event/repository/EventTypeRepository.java
@@ -1,6 +1,6 @@
-package com.softeer.podo.admin.repository;
+package com.softeer.podo.event.repository;
 
-import com.softeer.podo.admin.model.entity.EventType;
+import com.softeer.podo.event.model.entity.EventType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventTypeRepository extends JpaRepository<EventType, Long> {
