+++ src/main/java/com/softeer/podo/event/repository/EventWeightRepository.java
@@ -1,6 +1,6 @@
-package com.softeer.podo.admin.repository;
+package com.softeer.podo.event.repository;
 
-import com.softeer.podo.admin.model.entity.EventWeight;
+import com.softeer.podo.event.model.entity.EventWeight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventWeightRepository extends JpaRepository<EventWeight, Long> {
