+++ src/main/java/com/softeer/podo/event/repository/LotsUserRepository.java
@@ -1,6 +1,6 @@
-package com.softeer.podo.admin.repository;
+package com.softeer.podo.event.repository;
 
-import com.softeer.podo.admin.model.entity.LotsUser;
+import com.softeer.podo.event.model.entity.LotsUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
