+++ src/test/java/com/softeer/podo/admin/service/AdminServiceTest.java
@@ -5,13 +5,15 @@ import com.softeer.podo.admin.model.dto.request.EventConfigRequestDto;
 import com.softeer.podo.admin.model.dto.request.EventRewardConfigRequestDto;
 import com.softeer.podo.admin.model.dto.response.EventListResponseDto;
 import com.softeer.podo.admin.model.dto.response.EventRewardConfigResponseDto;
-import com.softeer.podo.admin.model.dto.user.ArrivalUserDto;
-import com.softeer.podo.admin.model.dto.user.ArrivalUserListDto;
-import com.softeer.podo.admin.model.dto.user.LotsUserDto;
-import com.softeer.podo.admin.model.dto.user.LotsUserListDto;
-import com.softeer.podo.admin.model.entity.*;
-import com.softeer.podo.admin.repository.*;
-import com.softeer.podo.event.model.entity.LotsComment;
+import com.softeer.podo.admin.model.dto.ArrivalUserDto;
+import com.softeer.podo.admin.model.dto.ArrivalUserListDto;
+import com.softeer.podo.admin.model.dto.LotsUserDto;
+import com.softeer.podo.admin.model.dto.LotsUserListDto;
+import com.softeer.podo.event.model.entity.*;
+import com.softeer.podo.event.repository.ArrivalUserRepository;
+import com.softeer.podo.event.repository.EventRepository;
+import com.softeer.podo.event.repository.EventRewardRepository;
+import com.softeer.podo.event.repository.LotsUserRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
